notebooks/requirements.txt
jupytext==1.14.5
plotly-express
pandas
nbformat>=4.2.0
python-dotenv